Added getUserToDo route and also fixed various bugs and completed the backend part

+++ src/controllers/deleteTodo.ts
@@ -1,6 +1,5 @@
 import { PrismaClient } from "@prisma/client";
 import { Request, Response, NextFunction } from "express";
-import { todo } from "node:test";
 
 const prisma = new PrismaClient();
 
@@ -18,9 +17,17 @@ export default async function deleteTodo(
       res.status(400).json({ error: "Todo ID is required" });
       return;
     }
+    // Check if todo with given id exists
+    const todo = await prisma.todo.findUnique({
+      where: { id: parseInt(id) },
+    });
 
-    await prisma.todo.delete({where:{id:parseInt(id)}});
-    res.status(204).json({message:"Successfully deleted the todo"});
+    if (!todo) {
+      return res.status(404).json({ error: "Todo not found" });
+    }
+
+    await prisma.todo.delete({ where: { id: parseInt(id) } });
+    res.status(200).json({ message: "Successfully deleted the todo" });
   } catch (error) {
     next(error);
   }
